backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
django-filter>=24.2
mysqlclient>=2.2
Pillow>=10.0
python-decouple>=3.8
